Reorganized the Housing components & page as well as making necessary changes on their imports if necessary.
