+++ employee-portal/src/components/housing/HousingDetails.tsx
